refactor: extracting heading into a custom component
