+++ src/components/TemplateMapping/TemplateMapping.jsx
@@ -6,7 +6,7 @@ import MappingPropertiesForm from "../TemplateMappingComponents/MappingPropertie
 import Mapper from "../TemplateMappingComponents/Mapper";
 import CustomButton from "../Styled/CustomButton";
 import CustomSpinner from "../Shared/CustomSpinner";
-import { t } from "i18next";
+import ContentHeading from "./ContentHeading";
 
 function TemplateMapping({ data, locale, mappingId }) {
   // --- STATE ---
@@ -93,34 +93,3 @@ function TemplateMapping({ data, locale, mappingId }) {
 }
 
 export default TemplateMapping;
-
-function ContentHeading({ templateMappingId, saveMapping }) {
-  const stickyStyle = {
-    position: "sticky",
-    top: "60px", // Replace with navbar height constant
-    zIndex: 1000,
-    backgroundColor: "white",
-    width: "100%",
-  };
-
-  return (
-    <div className="row content-heading" style={stickyStyle}>
-      <div className="col-md-12">
-        <h1 style={{ display: 'flex', justifyContent: 'space-between' }}>
-          <div>
-            <a href="/super_admin/template_mappings" className="btn btn-primary pull-left">{t('← All Mappings')}</a>
-          </div>
-          <span>{(templateMappingId ? `${t('Edit Mapping')} ${templateMappingId}` : t('Create Mapping'))}</span>
-          <div>
-            <div style={{ visibility: templateMappingId ? 'visible' : 'hidden' }}>
-              <CustomButton
-                title="💾 Save mapping"
-                handleClick={saveMapping}
-                buttonColor="orange" />
-            </div>
-          </div>
-        </h1>
-      </div>
-    </div>
-  );
-}
